Add constants package with special metaclass.
To use this add a class to constants.py and subclass `CONSTANTS`. Set class variables with an appropriate name and annotate for clarity.

Sections in `config.toml` that match the class name.upper() will override values of the set class attributes *IF THEY EXIST*.

For example:

Class `Test` in constants.py with attribute THING set to 1.
Section `TEST` in config.toml with option THING set to 2 and option OTHER set to 3.

`constants.Test.THING` will now = 2.
`constants.Test.OTHER` will not exist.

Once set constants can not be reset or deleted.